+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2021-01-30
+
 - Switch to godaemon to demonize [GH-15](https://github.com/mnencia/mchfuse/issues/17)
 - Store the version information inside the executable.
   The build version is printed in the log during the startup.
@@ -47,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/mnencia/mchfuse/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/mnencia/mchfuse/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/mnencia/mchfuse/releases/tag/v0.3.2
 [0.3.1]: https://github.com/mnencia/mchfuse/releases/tag/v0.3.1
 [0.3.0]: https://github.com/mnencia/mchfuse/releases/tag/v0.3.0
 [0.2.0]: https://github.com/mnencia/mchfuse/releases/tag/v0.2.0